backend/pyproject.toml
[project]
name = "family-learning-backend"
version = "0.1.0"
description = "Family Learning App - Backend"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy>=2.0",
    "alembic>=1.14",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.18",
    "jsonschema>=4.23",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.28",
    "pytest-asyncio>=0.25",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"
